Update main.py
Final code

+++ main.py
@@ -7,17 +7,18 @@ from langchain.chains import RetrievalQAWithSourcesChain
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain.document_loaders import UnstructuredURLLoader
 from langchain.embeddings import OpenAIEmbeddings
+# from langchain.embeddings import HuggingFaceEmbeddings
 from langchain.vectorstores import FAISS
 
 from dotenv import load_dotenv
 load_dotenv()  # take environment variables from .env (especially openai api key)
 
-st.title("StockSage: Equity Research Analyzer 📈")
+st.title("News Research Tool 📈")
 st.sidebar.title("News Article URLs")
 
 urls = []
 for i in range(3):
-    url = st.sidebar.text_input(f"URL {i+1}")
+    url= st.sidebar.text_input(f"URL {i+1}")
     urls.append(url)
 
 process_url_clicked = st.sidebar.button("Process URLs")
@@ -29,24 +30,32 @@ llm = OpenAI(temperature=0.9, max_tokens=500)
 if process_url_clicked:
     # load data
     loader = UnstructuredURLLoader(urls=urls)
-    main_placeholder.text("Data Loading...Started...✅✅✅")
+    main_placeholder.text("Data Loading >>> Started >>> ✅✅✅")
     data = loader.load()
+    print(len(data))
     # split data
     text_splitter = RecursiveCharacterTextSplitter(
         separators=['\n\n', '\n', '.', ','],
         chunk_size=1000
     )
-    main_placeholder.text("Text Splitter...Started...✅✅✅")
+    main_placeholder.text("Text Splitter >>> Started >>> ✅✅✅")
     docs = text_splitter.split_documents(data)
-    # create embeddings and save it to FAISS index
-    embeddings = OpenAIEmbeddings()
-    vectorstore_openai = FAISS.from_documents(docs, embeddings)
-    main_placeholder.text("Embedding Vector Started Building...✅✅✅")
+    
+    if os.path.exists(file_path):
+        with open(file_path, "rb") as f:
+            vectorstore_openai = pickle.load(f)
+    else:
+        # create embeddings and save it to FAISS index
+        embeddings = OpenAIEmbeddings()
+        # Example with a lightweight model
+        # embeddings = HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")
+        
+        # Process and save the Faiss index into a pickle file
+        vectorstore_openai = FAISS.from_documents(docs, embeddings)
+        with open(file_path, "wb") as f:
+            pickle.dump(vectorstore_openai, f)
+    main_placeholder.text("Embedding Vector Creation >>> ✅✅✅")
     time.sleep(2)
-
-    # Save the FAISS index to a pickle file
-    with open(file_path, "wb") as f:
-        pickle.dump(vectorstore_openai, f)
 
 query = main_placeholder.text_input("Question: ")
 if query:
@@ -67,6 +76,22 @@ if query:
                 for source in sources_list:
                     st.write(source)
 
+# --------------------------------------------------
 
 
+# from langchain.document_loaders import TextLoader
+# import requests
 
+# texts = []
+# for url in urls:
+#     try:
+#         response = requests.get(url)
+#         response.raise_for_status()
+#         texts.append(response.text)  # Fetch raw HTML
+#     except Exception as e:
+#         print(f"Error fetching {url}: {e}")
+
+# # Load documents from fetched raw HTML
+# loader = TextLoader(documents=texts)
+# documents = loader.load()
+# print(f"Loaded {len(documents)} documents")
